office-booking/requirements.txt
git+https://github.com/hunkim/streamlit-google-oauth
protobuf<=3.20.x 
streamlit
httpx-oauth
python-dotenv
typing_extensions
gsheetsdb